+++ src/main/java/io/swagger/model/DepositOrWithdrawRequestDTO.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.math.BigDecimal;
+
 import org.springframework.validation.annotation.Validated;
 import javax.validation.Valid;
 import javax.validation.constraints.*;
@@ -18,12 +18,12 @@ import javax.validation.constraints.*;
 
 public class DepositOrWithdrawRequestDTO   {
   @JsonProperty("amount")
-  private BigDecimal amount = null;
+  private Double amount = null;
 
   @JsonProperty("iban")
   private String iban = null;
 
-  public DepositOrWithdrawRequestDTO amount(BigDecimal amount) {
+  public DepositOrWithdrawRequestDTO amount(Double amount) {
     this.amount = amount;
     return this;
   }
@@ -36,11 +36,11 @@ public class DepositOrWithdrawRequestDTO   {
       @NotNull
 
     @Valid
-    public BigDecimal getAmount() {
+    public Double getAmount() {
     return amount;
   }
 
-  public void setAmount(BigDecimal amount) {
+  public void setAmount(Double amount) {
     this.amount = amount;
   }
