+++ src/main/java/io/swagger/model/Transaction.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.math.BigDecimal;
+
 import org.springframework.validation.annotation.Validated;
 import javax.validation.Valid;
 import javax.validation.constraints.*;
@@ -25,7 +25,7 @@ public class Transaction   {
   private String ibanTo = null;
 
   @JsonProperty("amount")
-  private BigDecimal amount = null;
+  private Double amount = null;
 
   @JsonProperty("userPerforming")
   private Integer userPerforming = null;
@@ -105,7 +105,7 @@ public class Transaction   {
     this.ibanTo = ibanTo;
   }
 
-  public Transaction amount(BigDecimal amount) {
+  public Transaction amount(Double amount) {
     this.amount = amount;
     return this;
   }
@@ -118,11 +118,11 @@ public class Transaction   {
       @NotNull
 
     @Valid
-    public BigDecimal getAmount() {
+    public Double getAmount() {
     return amount;
   }
 
-  public void setAmount(BigDecimal amount) {
+  public void setAmount(Double amount) {
     this.amount = amount;
   }
